+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and includes an additional section for migration notes.
 - *ORCA-554*, *ORCA-561*, *ORCA-579*, *ORCA-581*
   - GraphQL image, service, and Load Balancer will now be deployed by TF.
   - *ORCA-420* Added Internal Reconcile Report Mismatch functionality to GraphQL.
+  - *ORCA-556* Added Internal Reconcile Report Phantom functionality to GraphQL.
   - *ORCA-592* GraphQL logs are json structures, and can thus be queried in CloudWatch.
 - *ORCA-597*
   - Server access logging is now enabled for graphql application load balancer.